gateway/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==6.33.2